fix(p3/script): check argocd existence

+++ p3/scripts/init.sh
@@ -25,9 +25,15 @@ fi
 
 # Install Argo CD CLI
 echo -e "\033[34m$(date +%T) [INFO] Installing Argo CD\033[0m"
-curl -sSL -o argocd-linux-amd64 https://github.com/argoproj/argo-cd/releases/latest/download/argocd-linux-amd64
-sudo install -m 555 argocd-linux-amd64 /usr/local/bin/argocd
-rm argocd-linux-amd64
+if command -v argocd >/dev/null 2>&1; then
+  echo "ArgoCD already exists."
+else {
+  curl -sSL -o argocd-linux-amd64 https://github.com/argoproj/argo-cd/releases/latest/download/argocd-linux-amd64
+  sudo install -m 555 argocd-linux-amd64 /usr/local/bin/argocd
+  rm argocd-linux-amd64
+}
+fi
+
 
 # Install Argo CD setup
 echo -e "\033[34m$(date +%T) [INFO] Argo CD Setup\033[0m"
